requirements.txt
flask
twilio
requests
openai
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
